+++ benchmark.sh
@@ -19,6 +19,7 @@ hyperfine --warmup 3 -m 5 --export-markdown single_char.md --show-output \
 
 # Multi character delimiter tests, create a file with three spaces as delmiter
 hck -d, -f1- -D '   ' -o ./hyper_data_multichar.txt ./hyper_data.txt
+sed -i 's/# label/#label/' ./hyper_data_multichar.txt
 
 # Notes:
 #   - xsv gets messed up by the first commented line in the section, so tail skips it, the point is to run these commands as we would in the wild
@@ -27,8 +28,12 @@ hyperfine --warmup 3 -m 5 --export-markdown multi_char.md --show-output \
     "hck -Ld'   ' -f1,8,19 --no-mmap ./hyper_data_multichar.txt > /dev/null" \
     "hck -d'[[:space:]]+' -f1,8,19 ./hyper_data_multichar.txt > /dev/null" \
     "hck -d'[[:space:]]+' --no-mmap -f1,8,19 ./hyper_data_multichar.txt > /dev/null" \
+    "hck -d'[[:space:]]' -f1,8,19 -g ./hyper_data_multichar.txt > /dev/null" \
+    "hck -d'[[:space:]]' --no-mmap -f1,8,19 -g ./hyper_data_multichar.txt > /dev/null" \
     "hck -d'\s+' -f1,8,19 ./hyper_data_multichar.txt > /dev/null" \
     "hck -d'\s+' -f1,8,19 --no-mmap ./hyper_data_multichar.txt > /dev/null" \
+    "hck -d'\s' -f1,8,19 -g ./hyper_data_multichar.txt > /dev/null" \
+    "hck -d'\s' -f1,8,19 -g --no-mmap ./hyper_data_multichar.txt > /dev/null" \
     "choose -f '   ' -i ./hyper_data_multichar.txt 0 7 18  > /dev/null" \
     "choose -f '[[:space:]]' -i ./hyper_data_multichar.txt 0 7 18  > /dev/null" \
     "choose -f '\s' -i ./hyper_data_multichar.txt 0 7 18  > /dev/null" \
@@ -36,7 +41,7 @@ hyperfine --warmup 3 -m 5 --export-markdown multi_char.md --show-output \
     "awk -F'   ' '{print \$1, \$8, \$19}' ./hyper_data_multichar.txt > /dev/null" \
     "awk -F'[:space:]+' '{print \$1, \$8, \$19}' ./hyper_data_multichar.txt > /dev/null" \
     "< ./hyper_data_multichar.txt tr -s ' ' | cut -d ' ' -f1,8,19 > /dev/null" \
-    "< ./hyper_data_multichar.txt tr -s ' ' | tail -n+2 | xsv select -d ' ' 1,8,19 --no-headers > /dev/null" \
+    "< ./hyper_data_multichar.txt tr -s ' ' | xsv select -d ' ' 1,8,19 --no-headers > /dev/null" \
     "< ./hyper_data_multichar.txt tr -s ' ' | hck -Ld' ' -f1,8,19 > /dev/null" \
     "< ./hyper_data_multichar.txt tr -s ' ' | tsv-select -d ' ' -f 1,8,19 > /dev/null"