+++ src/lib/line_parser.rs
@@ -49,7 +49,7 @@ impl<'a> LineParser<'a> for SubStrLineParser<'a> {
             for _ in 0..=high - low {
                 match parts.next() {
                     Some(part) => {
-                        // Guaranteed to be in range since staging is created based on field pos anyways
+                        // Guaranteed to be in range since shuffler is created based on field pos anyways
                         if let Some(reshuffled_range) = shuffler.get_mut(pos) {
                             reshuffled_range.push(part)
                         }
@@ -66,13 +66,16 @@ impl<'a> LineParser<'a> for SubStrLineParser<'a> {
 pub struct RegexLineParser<'a> {
     field_ranges: &'a [FieldRange],
     delimiter: &'a Regex,
+    // Controls whether or not to consume "empty" delimiters so that you don't need to write "\s+", only "\s"
+    greedy: bool,
 }
 
 impl<'a> RegexLineParser<'a> {
-    pub fn new(field_ranges: &'a [FieldRange], delimiter: &'a Regex) -> Self {
+    pub fn new(field_ranges: &'a [FieldRange], delimiter: &'a Regex, greedy: bool) -> Self {
         Self {
             field_ranges,
             delimiter,
+            greedy,
         }
     }
 }
@@ -82,7 +85,16 @@ impl<'a> LineParser<'a> for RegexLineParser<'a> {
     where
         'a: 'b,
     {
-        let mut parts = self.delimiter.split(line).peekable();
+        let mut parts: Box<dyn Iterator<Item = _>> = if self.greedy {
+            Box::new(
+                self.delimiter
+                    .split(line)
+                    .filter(|s| !s.is_empty())
+                    .peekable(),
+            )
+        } else {
+            Box::new(self.delimiter.split(line).peekable())
+        };
         let mut iterator_index = 0;
 
         // Iterate over our ranges and write any fields that are contained by them.
@@ -101,7 +113,7 @@ impl<'a> LineParser<'a> for RegexLineParser<'a> {
             for _ in 0..=high - low {
                 match parts.next() {
                     Some(part) => {
-                        // Guaranteed to be in range since staging is created based on field pos anyways
+                        // Guaranteed to be in range since shuffler is created based on field pos anyways
                         if let Some(reshuffled_range) = shuffler.get_mut(pos) {
                             reshuffled_range.push(part)
                         } else {